+++ server/config/index.js
@@ -2,7 +2,8 @@
 
 module.exports = {
   default: ({ env }) => ({
-    imagePath: env("CW_IMAGE_PATH") ?? ".tokens",
+    assetPath: env("CW_ASSET_PATH") ?? ".tokens",
+    verificationMessage: "sign in",
     infuraProjectId: env("infura_project_id") ?? "INFURA PROJECT ID"
   }),
   validator() {},