+++ CHANGELOG.md
@@ -9,5 +9,6 @@ v0.1.1
 * Allow abi and metadata to be returned in API calls
 
 v0.2.0
+* Added ability to open managed wallet
 * Removed Network entity and associated references
 * Fixed bugs in token lifecycle and contract service