+++ CHANGELOG.md
@@ -13,3 +13,4 @@ v0.2.0
 * Removed Network entity and associated references
 * Fixed bugs in token lifecycle and contract service
 * Added mint token*
+* Added getOrCreateUserWallet